rename doctor modal autocomplete js
