+++ app/javascript/packs/application.js
@@ -18,4 +18,4 @@ import 'bootstrap';
 
 import '../src/application.scss';
 import './flashMessages';
-import './modal_doctor_autocomplete';
+import './modalDoctorAutocomplete';
