+++ app/javascript/packs/modalDoctorAutocomplete.js
